Script updates
Added a couple of new restart scripts, moved them all to the System directory. Updated the install script to symlink to the Nautilus scripts. Removed the Move Unity script as the issue I was having no longer exists in 11.10, but more importantly, I moved off Unity to GnomeShell so even if it still existed... yeah.