[zsh] adds rvm, alias and shebangs
OS X is stupid and ships with a way old version of Ruby. RVM is the only
solution but it slows the prompt down considerably so I made it OS X
only. Added an alias for tmuxinator as the brew installed version
doesn't alias to mux out of the box. Added shebangs to files to they are
properly syntax highlighted.