[zsh] keybindings and tab completion improvements
added the highlight when hitting tab and some of the keybindings I was
missing. interestingly enough, I've come to realize that a lot of the
keybindings you have to set in zsh appear to be default in bash.
starting to make a migration back to bash a logical progression.

need to go through and figure out which options that I'm setting or
unsetting are default options and trim down to just what I need.