[zsh] dropped a bunch of aliases I never use
Trying to get closer to a vanilla setup so I'm not crippled when I am accessing
new servers and other people's machines and such. So many times I am on a
machine that has vim installed but NOT my v alias. It's frustrating and I feel
like it wastes a ton more time than it saves.