Added TabNine, cleaned up some of the installer
Installed was running as `sh` and a bunch of crap was lit up as being wrong in
POSIX mode, so just changed it over to run via `bash`. Added TabNine to the vim
install and added the appropriate `vimrc` stuff for that. Also dropped some old
color toggling shit I'm not using anymore.