Whelp, couldn't convince myself to switch to tmux
Did some soul searching and I really couldn't find any compelling reason to
switch to tmux. Sure it's modern and has a ton of cool features, but said
features actually don't interest me. Couple that with the fact that I was
planning to reconfigure tmux a bit to use screen's same mapping key, just made
me think there's no reason to switch if all I'm going to be doing is trying to
will one to work like the other.