Increased scroll back and turned status line on
Finding that I get a bit lost when I have 6+ windows open, so added back the
hard status line. Wasn't entirely sure if I wanted to make it fancy, or just
stick with the absolutely boring stock colors, so left a more advanced snippet
in there in case I wanted to use it later.