feat(zsh): drop "vim prompt"
I once was lost, but now I'm found. Original issue where I would forget
if I was inside of a shell inside of `vim` has been pretty much
nullified as I rarely use terminals inside of `nvim` and when I do, it's
quite obvious that I'm inside of a shell that's inside of `vim` and not
obscured like it was in the `:sh[ell]` days.