Spring clean up
Quite a few updates to leverage GNU `stow` and reduce the manual efforts
in the install script. Localized my `[n]vim` color scheme of choice
since it hasn't been updated in years. Also updated the installer to
pull from remote instead of completely removing itself if it's already
present.

Contemplating moving to using some plugin managers for `zsh` and
`[n]vim` to help reduce the install size a bit more, and to provide a
bit more flexibility for anybody else running this.