Boomeranging back to Neovim
Been playing with new color schemes as I've switched to Alacritty. Ran
into a scenario where there was a color scheme for Neovim and not Vim.
New color scheme is solid, so my previous color scheme issues are now
null and void.

Instead of juggling around with a shared configuration, I've committed
to a separate configuration for Vim and Neovim.